+++ Controllers/BackofficeFacilitiesCategoryDeleteController.class.php
@@ -1,5 +1,4 @@
 <?php
-require_once 'Database/UserRepository.class.php';
 require_once 'Database/CategoryRepository.class.php';
 require_once 'Middlewares/SessionManager.class.php';
 require_once 'Models/User.class.php';
@@ -10,23 +9,25 @@ require_once 'Views/Html404.class.php';
 
 class BackofficeFacilitiesCategoryDeleteController
 {
-    protected $user_repository;
     protected $category_repository;
-
 
     public function __construct()
     {
-        $this->user_repository = new UserRepository();
         $this->category_repository = new CategoryRepository();
     }
 
-    public function http_get(array &$params): IView
+    public function http_post(array &$params): IView
     {
-        if (isset($params['category'])) {
+        $user = SessionManager::get_user();
+        if (User::is_empty($user))
+        {
+            return new HttpRedirectView('/backoffice');
+        }
 
+        if (isset($params['category']))
+        {
             $id = intval($params['facility']);
             $this->category_repository->delete_category($id);
-
         }
 
         return new HttpRedirectView('/backoffice/facilities/categories');
