+++ Controllers/BackofficeGuestDeleteController.class.php
@@ -1,5 +1,4 @@
 <?php
-require_once 'Database/UserRepository.class.php';
 require_once 'Database/GuestRepository.class.php';
 require_once 'Middlewares/SessionManager.class.php';
 require_once 'Models/User.class.php';
@@ -10,23 +9,25 @@ require_once 'Views/Html404.class.php';
 
 class BackofficeGuestDeleteController
 {
-    protected $user_repository;
     protected $guest_repository;
 
     public function __construct()
     {
-        $this->user_repository = new UserRepository();
         $this->guest_repository = new GuestRepository();
     }
 
     public function http_post(array &$params): IView
     {
-        if (isset($params['guest'])) {
-            $user = SessionManager::get_user();
+        $user = SessionManager::get_user();
+        if (User::is_empty($user))
+        {
+            return new HttpRedirectView('/backoffice');
+        }
 
+        if (isset($params['guest']))
+        {
             $id = intval($params['guest']);
             $this->guest_repository->remove_by_id($id);
-
         }
 
         return new HttpRedirectView('/backoffice/guests');
