+++ Controllers/BackofficeHotelDeleteController.class.php
@@ -1,5 +1,4 @@
 <?php
-require_once 'Database/UserRepository.class.php';
 require_once 'Database/HotelRepository.class.php';
 require_once 'Middlewares/SessionManager.class.php';
 require_once 'Models/User.class.php';
@@ -10,22 +9,25 @@ require_once 'Views/Html404.class.php';
 
 class BackofficeHotelDeleteController
 {
-    protected $user_repository;
     protected $hotel_repository;
 
     public function __construct()
     {
-        $this->user_repository = new UserRepository();
         $this->hotel_repository = new HotelRepository();
     }
 
-    public function http_get(array &$params): IView
+    public function http_post(array &$params): IView
     {
-        if (isset($params['hotel'])) {
+        $user = SessionManager::get_user();
+        if (User::is_empty($user))
+        {
+            return new HttpRedirectView('/backoffice');
+        }
 
+        if (isset($params['hotel']))
+        {
             $id = intval($params['hotel']);
             $this->hotel_repository->delete_hotel($id);
-
         }
 
         return new HttpRedirectView('/backoffice/hotels');
